+++ src/main/java/ar/com/ddsutn/TP1Macowins/Zapato.java
@@ -2,9 +2,9 @@ package ar.com.ddsutn.TP1Macowins;
 
 public class Zapato extends Prenda {
 
-	private int talle;
+	private double talle;
 
-	public Zapato(Origen origen, Marca marca, int talle) {
+	public Zapato(Origen origen, Marca marca, double talle) {
 		super(origen, marca);
 		this.talle = talle;
 	}
